+++ src/Http/Controllers/CouponController.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Vanilo\Admin\Http\Controllers;
+
+class CouponController
+{
+}